Create plisi-2014-10
Plisi Albanian keyboard layout code for Linux. The code should be pasted into the "al" file in "/usr/share/x11/xkb/symbols". This Linux version of the layout is not identical to the Windows and OSX versions as it is missing a second group of additional characters. Please help complete the layout.